+++ src/Headers/Client.hpp
@@ -8,14 +8,14 @@
 class Client
 {
 private:
-	double x;
-	double y;
+	long double x;
+	long double y;
 	std::vector<Item *> items;
 
 public:
 	int id;
 	Client(const std::vector<std::string> &values);
-	Client(int id, double x = 0, double y = 0);
+	Client(int id, long double x = 0, long double y = 0);
 	~Client();
 	void addItem(Item *item);
 	void print();
