+++ src/Headers/Item.hpp
@@ -11,17 +11,17 @@ private:
 public:
 	Vehicle *vehicle = nullptr;
 	Point destination;
-	double weight;
+	long double weight;
 	int type;
 	int clientId;
 	int id;
 	Item(const std::vector<std::string> &values);
-	Item(int id, int clientId, int type, double weight);
+	Item(int id, int clientId, int type, long double weight);
 	~Item();
 	void setVehicle(Vehicle *vehicle);
 	void setDestination(Point *point);
-	double distanceTo(Point *point);
-	void setWeight(double weight);
+	long double distanceTo(Point *point);
+	void setWeight(long double weight);
 	bool wasAttended();
 	void reset();
 };
