+++ src/tests/Vehicle_test.cpp
@@ -4,7 +4,7 @@
 
 TEST_CASE("Build from value list")
 {
-    // int id, int type, double capacity,  int carrierId, double costPerKm, double minimumCapacity, double additionalForMultipleClients, max distace
+    // int id, int type, long double capacity,  int carrierId, long double costPerKm, long double minimumCapacity, long double additionalForMultipleClients, max distace
     std::vector<std::string> values = {"1", "3", "0.1", "2", "0.2", "0.3", "0.4", "0.5"};
     Vehicle vehicle = Vehicle(values);
     REQUIRE(vehicle.id == 1);
@@ -25,16 +25,16 @@ TEST_CASE("Reset")
 
 TEST_CASE("Calculate item cost delta")
 {
-    double additionalForMultipleClients = 1;
-    double costPerKmPerWeight = 1;
-    double itemWeight = 1;
+    long double additionalForMultipleClients = 1;
+    long double costPerKmPerWeight = 1;
+    long double itemWeight = 1;
     int minimumCapacity = 1;
     Point *origin = new Point(0, 0);
     Point *destination = new Point(0, 2);
     Vehicle vehicle = Vehicle(1, 1, 1, 100, costPerKmPerWeight, minimumCapacity, additionalForMultipleClients, 10);
     Item item = Item(1, 1, 1, itemWeight);
     item.setDestination(destination);
-    double item1TripCost = costPerKmPerWeight * itemWeight * origin->distanceTo(destination);
+    long double item1TripCost = costPerKmPerWeight * itemWeight * origin->distanceTo(destination);
 
     SECTION("Single item")
     {
@@ -46,38 +46,38 @@ TEST_CASE("Calculate item cost delta")
         vehicle.setMinimumCapacity(minimumCapacity);
         itemWeight = minimumCapacity - 0.5;
         item.setWeight(itemWeight);
-        double deadfreightCost = costPerKmPerWeight * (minimumCapacity - itemWeight) * origin->distanceTo(destination);
-        double expectedCost = costPerKmPerWeight * (minimumCapacity - itemWeight) * origin->distanceTo(destination) + deadfreightCost;
+        long double deadfreightCost = costPerKmPerWeight * (minimumCapacity - itemWeight) * origin->distanceTo(destination);
+        long double expectedCost = costPerKmPerWeight * (minimumCapacity - itemWeight) * origin->distanceTo(destination) + deadfreightCost;
         REQUIRE(vehicle.calculateTripCostDelta(&item) == expectedCost);
     }
 
     SECTION("Multiple items over deadfreight")
     {
         vehicle.take(&item);
-        double secondItemWeight = 2;
+        long double secondItemWeight = 2;
         destination = new Point(0, 3);
         Item item2 = Item(1, 2, 1, secondItemWeight);
         item2.setDestination(destination);
-        double deadfreightCost = costPerKmPerWeight * (minimumCapacity - itemWeight) * origin->distanceTo(destination);
-        double expectedCost = costPerKmPerWeight * (secondItemWeight)*origin->distanceTo(destination) + additionalForMultipleClients + item1TripCost;
+        long double deadfreightCost = costPerKmPerWeight * (minimumCapacity - itemWeight) * origin->distanceTo(destination);
+        long double expectedCost = costPerKmPerWeight * (secondItemWeight)*origin->distanceTo(destination) + additionalForMultipleClients + item1TripCost;
         REQUIRE(vehicle.calculateTripCostDelta(&item2) == expectedCost - item1TripCost + deadfreightCost);
     }
 
     SECTION("Multiple items under deadfreight")
     {
-        double firstItemWeight = 0.5, secondItemWeight = 0.3;
+        long double firstItemWeight = 0.5, secondItemWeight = 0.3;
         int deadfreight = 1;
         vehicle.setMinimumCapacity(deadfreight);
         item.setWeight(firstItemWeight);
-        double firstTripCost = vehicle.calculateTripCostDelta(&item);
-        double deadfreightCost = costPerKmPerWeight * (minimumCapacity - firstItemWeight) * origin->distanceTo(destination);
+        long double firstTripCost = vehicle.calculateTripCostDelta(&item);
+        long double deadfreightCost = costPerKmPerWeight * (minimumCapacity - firstItemWeight) * origin->distanceTo(destination);
 
         SECTION("Second item still have deadfreight")
         {
             Item item2 = Item(1, 2, 1, secondItemWeight);
             destination = new Point(0, 3);
             item2.setDestination(destination);
-            double secondTripCost = origin->distanceTo(destination) * costPerKmPerWeight * (secondItemWeight);
+            long double secondTripCost = origin->distanceTo(destination) * costPerKmPerWeight * (secondItemWeight);
             secondTripCost += origin->distanceTo(destination) * costPerKmPerWeight * (deadfreight - firstItemWeight - secondItemWeight);
             secondTripCost += additionalForMultipleClients;
 
@@ -90,7 +90,7 @@ TEST_CASE("Calculate item cost delta")
             Item item2 = Item(1, 2, 1, secondItemWeight);
             destination = new Point(0, 3);
             item2.setDestination(destination);
-            double secondTripCost = origin->distanceTo(destination) * costPerKmPerWeight * (secondItemWeight);
+            long double secondTripCost = origin->distanceTo(destination) * costPerKmPerWeight * (secondItemWeight);
             secondTripCost += additionalForMultipleClients;
 
             vehicle.take(&item);
